feat(SizeUtils): allow MediaQueryData override
This change introduces the ability to override the `mediaQueryData` used by `SizeUtils`.

A backing field `_mediaQueryDataOverride` and a setter for `mediaQueryData` are added. This allows widgets with access to `BuildContext` to provide their own `MediaQueryData`, enabling more flexible responsive UI adjustments.

The `mediaQueryData` getter now prioritizes the override if it's set.